Handle some game logic

+++ src/com/jgegroup/pacman/objects/Consumables.java
@@ -2,13 +2,13 @@ package jgegroup.pacman.objects;
 
 public abstract class Consumables extends GameObjects{
     protected int score;
-    public Consumables(double x, double y) {
+    public Consumables(int x, int y) {
         this(x,y,0);
     }
     public Consumables(Position position){
         this(position,0);
     }
-    public Consumables(double x, double y, int score) {
+    public Consumables(int x, int y, int score) {
         super(x,y);
         this.score=score;
     }
